+++ app/src/main/java/ee/taltech/kamatt/sportsmap/db/repository/GpsLocationRepository.kt
@@ -70,7 +70,7 @@ class GpsLocationRepository(val context: Context) {
                     cursor.getDouble(cursor.getColumnIndex("accuracy")),
                     cursor.getDouble(cursor.getColumnIndex("altitude")),
                     cursor.getDouble(cursor.getColumnIndex("verticalAccuracy")),
-                    cursor.getLong(cursor.getColumnIndex("gpsSessionId")),
+                    cursor.getInt(cursor.getColumnIndex("gpsSessionId")),
                     cursor.getString(cursor.getColumnIndex("gpsLocationTypeId")),
                     cursor.getInt(cursor.getColumnIndex("appUserId"))
 
@@ -106,7 +106,7 @@ class GpsLocationRepository(val context: Context) {
                     cursor.getDouble(cursor.getColumnIndex("accuracy")),
                     cursor.getDouble(cursor.getColumnIndex("altitude")),
                     cursor.getDouble(cursor.getColumnIndex("verticalAccuracy")),
-                    cursor.getLong(cursor.getColumnIndex("gpsSessionId")),
+                    cursor.getInt(cursor.getColumnIndex("gpsSessionId")),
                     cursor.getString(cursor.getColumnIndex("gpsLocationTypeId")),
                     cursor.getInt(cursor.getColumnIndex("appUserId"))
                 )
